requirements.txt
Flask== 2.2.2
Flask-Cors== 3.0.10
pymongo== 4.2.0
python-dotenv== 0.20.0
dnspython==2.2.1

#pip install -r requirements.txt